+++ examples/pptxgenjs-demo.js
@@ -1,7 +1,7 @@
 /**
 * NAME: pptxgenjs-demo.js
 * AUTH: Brent Ely (https://github.com/gitbrent/)
-* DATE: Apr 29, 2017
+* DATE: Aug 15, 2017
 * DESC: Common test/demo slides for all library features
 * DEPS: Loaded by `pptxgenjs-demo.js` and `nodejs-demo.js`
 */
@@ -64,12 +64,18 @@ function runEveryTest() {
 
 function execGenSlidesFuncs(type) {
 	// STEP 1: Instantiate new PptxGenJS object
+	var pptx;
 	if ( NODEJS ) {
-		//var pptx = require('../dist/pptxgen.js'); // for LOCAL TESTING
-		var pptx = require("pptxgenjs");
+		var fs = require('fs');
+		if (fs.existsSync('../dist/pptxgen.js')) {
+			pptx = require('../dist/pptxgen.js'); // for LOCAL TESTING
+		}
+		else {
+			pptx = require("pptxgenjs");
+		}
 	}
 	else {
-		var pptx = new PptxGenJS();
+		pptx = new PptxGenJS();
 	}
 
 	pptx.setLayout('LAYOUT_WIDE');
@@ -499,7 +505,7 @@ function genSlides_Chart(pptx) {
 		arrDataLineStat.push( tmpObjUnk );
 	}
 
-	// SLIDE 1: Bar Chart -----------------------------------------------------------------
+	// SLIDE 1: Bar Chart ------------------------------------------------------------------
 	{
 		var slide = pptx.addNewSlide();
 		slide.addTable( [ [{ text:'Chart Examples: Bar Chart', options:gOptsTitle }] ], { x:0.5, y:0.13, w:12.5 } );
@@ -621,7 +627,145 @@ function genSlides_Chart(pptx) {
 		slide.addChart( pptx.charts.BAR, arrDataHighVals, optsChartBar4 );
 	}
 
-	// SLIDE 2: Stacked Bar Chart ---------------------------------------------------------
+	// SLIDE 2: Bar Chart Grid/Axis Options ------------------------------------------------
+	{
+		var slide = pptx.addNewSlide();
+		slide.addTable( [ [{ text:'Chart Examples: Bar Chart Grid/Axis Options', options:gOptsTitle }] ], { x:0.5, y:0.13, w:12.5 } );
+
+		var arrDataRegions = [
+			{
+				name  : 'Region 1',
+				labels: ['May', 'June', 'July', 'August'],
+				values: [26, 53, 100, 75]
+			},
+			{
+				name  : 'Region 2',
+				labels: ['May', 'June', 'July', 'August'],
+				values: [43.5, 70.3, 90.1, 80.05]
+			}
+		];
+		var arrDataHighVals = [
+			{
+				name  : 'California',
+				labels: ['Apartment', 'Townhome', 'Duplex', 'House', 'Big House'],
+				values: [2000, 2800, 3200, 4000, 5000]
+			},
+			{
+				name  : 'Texas',
+				labels: ['Apartment', 'Townhome', 'Duplex', 'House', 'Big House'],
+				values: [1400, 2000, 2500, 3000, 3800]
+			}
+		];
+
+		// TOP-LEFT: H/bar
+		var optsChartBar1 = { x:0.5, y:0.6, w:6.0, h:3.0,
+			barDir: 'bar',
+			fill: 'F1F1F1',
+
+			catAxisLabelColor   : 'CC0000',
+			catAxisLabelFontFace: 'Helvetica Neue',
+			catAxisLabelFontSize: 14,
+
+			catGridLine: 'none',
+			catAxisHidden: true,
+			valGridLine: { color: "cc6699", style: "dash", size: 1 },
+
+			showLegend   : true,
+			title        : 'No CatAxis, ValGridLine=dash',
+			titleColor   : 'a9a9a9',
+			titleFontFace: 'Helvetica Neue',
+			titleFontSize: 14,
+			showTitle    : true
+		};
+		slide.addChart( pptx.charts.BAR, arrDataRegions, optsChartBar1 );
+
+		// TOP-RIGHT: V/col
+		var optsChartBar2 = { x:7.0, y:0.6, w:6.0, h:3.0,
+			barDir: 'col',
+			fill: 'E1F1FF',
+
+			dataBorder         : { pt:'1', color:'F1F1F1' },
+			dataLabelColor     : '696969',
+			dataLabelFontFace  : 'Arial',
+			dataLabelFontSize  : 11,
+			dataLabelPosition  : 'outEnd',
+			dataLabelFormatCode: '#.0',
+			showValue          : true,
+
+			catAxisHidden: true,
+			catGridLine  : 'none',
+			valAxisHidden: true,
+			valGridLine  : 'none',
+
+			showLegend: true,
+			legendPos : 'b',
+			showTitle : false
+		};
+		slide.addChart( pptx.charts.BAR, arrDataRegions, optsChartBar2 );
+
+		// BTM-LEFT: H/bar - TITLE and LEGEND
+		slide.addText( '.', { x:0.5, y:3.8, w:6.0, h:3.5, fill:'F1F1F1', color:'F1F1F1'} );
+		var optsChartBar3 = { x:0.5, y:3.8, w:6.0, h:3.5,
+			barDir     : 'bar',
+
+			border: { pt:'3', color:'CF0909' },
+			fill: 'F1C1C1',
+
+			catAxisLabelColor   : 'CC0000',
+			catAxisLabelFontFace: 'Helvetica Neue',
+			catAxisLabelFontSize: 14,
+			catAxisOrientation  : 'maxMin',
+			catAxisTitle: "Housing Type",
+			catAxisTitleColor: "428442",
+			catAxisTitleFontSize: 14,
+			showCatAxisTitle: true,
+
+			valGridLine: 'none',
+			valAxisHidden: true,
+			catGridLine: { color: "cc6699", style: "dash", size: 1 },
+
+			titleColor   : '33CF22',
+			titleFontFace: 'Helvetica Neue',
+			titleFontSize: 16,
+
+			showTitle : true,
+			title: 'Sales by Region'
+		};
+		slide.addChart( pptx.charts.BAR, arrDataHighVals, optsChartBar3 );
+
+		// BTM-RIGHT: V/col - TITLE and LEGEND
+		slide.addText( '.', { x:7.0, y:3.8, w:6.0, h:3.5, fill:'F1F1F1', color:'F1F1F1'} );
+		var optsChartBar4 = { x:7.0, y:3.8, w:6.0, h:3.5,
+			barDir: 'col',
+			barGapWidthPct: 25,
+			chartColors: ['0088CC', '99FFCC'],
+			chartColorsOpacity: 50,
+			valAxisMinVal: 1000,
+			valAxisMaxVal: 5000,
+
+			catAxisLabelColor   : '0000CC',
+			catAxisLabelFontFace: 'Times',
+			catAxisLabelFontSize: 11,
+			catAxisOrientation  : 'minMax',
+
+			dataBorder         : { pt:'1', color:'F1F1F1' },
+			dataLabelColor     : 'FFFFFF',
+			dataLabelFontFace  : 'Arial',
+			dataLabelFontSize  : 10,
+			dataLabelPosition  : 'ctr',
+			showValue          : true,
+
+			valAxisHidden      : true,
+			catAxisTitle       : 'Housing Type',
+			showCatAxisTitle   : true,
+
+			showLegend: false,
+			showTitle : false
+		};
+		slide.addChart( pptx.charts.BAR, arrDataHighVals, optsChartBar4 );
+	}
+
+	// SLIDE 3: Stacked Bar Chart ----------------------------------------------------------
 	{
 		var slide = pptx.addNewSlide();
 		slide.addTable( [ [{ text:'Chart Examples: Bar Chart: Stacked/PercentStacked', options:gOptsTitle }] ], { x:0.5, y:0.13, w:12.5 } );
@@ -690,27 +834,17 @@ function genSlides_Chart(pptx) {
 		};
 		slide.addChart( pptx.charts.BAR, arrDataRegions, optsChartBar2 );
 
-		// BTM-LEFT: H/bar - TITLE and LEGEND
+		// BTM-LEFT: H/bar - 100% layout without axis labels
 		var optsChartBar3 = { x:0.5, y:3.8, w:6.0, h:3.5,
-			barDir     : 'bar',
-			barGrouping: 'percentStacked',
-
-			border: { pt:'3', color:'CF0909' },
-			fill: 'F1C1C1',
-
-			catAxisLabelColor   : 'CC0000',
-			catAxisLabelFontFace: 'Helvetica Neue',
-			catAxisLabelFontSize: 14,
-			catAxisOrientation  : 'minMax',
-
-			titleColor   : '33CF22',
-			titleFontFace: 'Helvetica Neue',
-			titleFontSize: 16,
-
-			showTitle : true,
-			title: 'Sales by Region'
+			barDir       : 'bar',
+			barGrouping  : 'percentStacked',
+			dataBorder   : { pt:'1', color:'F1F1F1' },
+			catAxisHidden: true,
+			valAxisHidden: true,
+			showTitle    : false,
+			layout       : {x:0.1, y:0.1, w:1, h:1}
 		};
-		slide.addChart( pptx.charts.BAR, arrDataHighVals, optsChartBar3 );
+		slide.addChart( pptx.charts.BAR, arrDataRegions, optsChartBar3 );
 
 		// BTM-RIGHT: V/col - TITLE and LEGEND
 		slide.addText( '.', { x:7.0, y:3.8, w:6.0, h:3.5, fill:'F1F1F1', color:'F1F1F1'} );
@@ -726,14 +860,12 @@ function genSlides_Chart(pptx) {
 			chartColors: ['5DA5DA','FAA43A'],
 
 			showLegend: true,
-			legendPos :  't',
-			showTitle: true,
-			title    : 'Chart Title'
+			legendPos :  't'
 		};
 		slide.addChart( pptx.charts.BAR, arrDataHighVals, optsChartBar4 );
 	}
 
-	// SLIDE 3: Bar Chart - Lots of Bars --------------------------------------------------
+	// SLIDE 4: Bar Chart - Lots of Bars ---------------------------------------------------
 	{
 		var slide = pptx.addNewSlide();
 		slide.addTable( [ [{ text:'Chart Examples: Lots of Bars (>26 letters)', options:gOptsTitle }] ], { x:0.5, y:0.13, w:12.5 } );
@@ -749,17 +881,159 @@ function genSlides_Chart(pptx) {
 		var optsChart = {
 			x:0.5, y:0.5, w:'90%', h:'90%',
 			barDir: 'col',
-			title: 'Chart With >26 Cols'
+			title: 'Chart With >26 Cols',
+			showTitle: true,
+			titleFontSize: 20,
+			titleRotate: 10,
+			showCatAxisTitle: true,
+			catAxisTitle: "Letters",
+			catAxisTitleColor: "4286f4",
+			catAxisTitleFontSize: 14,
+
+			showValAxisTitle: true,
+			valAxisTitle: "Column Index",
+			valAxisTitleColor: "c11c13",
+			valAxisTitleFontSize: 16,
 		};
 
 		// TEST `getExcelColName()` to ensure Excel Column names are generated correctly above >26 chars/cols
 		slide.addChart(pptx.charts.BAR, arrDataHighVals, optsChart);
 	}
 
-	// SLIDE 4: Line Chart: Line Smoothing, Line Size, Symbol Size -------------------------------------------
+	// SLIDE 5: Bar Chart: Data Series Colors, majorUnits, and valAxisLabelFormatCode ------
 	{
 		var slide = pptx.addNewSlide();
-		slide.addTable( [ [{ text:'Chart Examples: Line Chart Smoothing, Line Size, Symbol Size', options:gOptsTitle }] ], { x:0.5, y:0.13, w:12.5 } );
+		slide.addTable( [ [{ text:'Chart Examples: Bar Colors, valAxisMajorUnit, v Format %', options:gOptsTitle }] ], { x:0.5, y:0.13, w:12.5 } );
+
+		// TOP-LEFT
+		slide.addChart(
+			pptx.charts.BAR,
+			[
+				{
+					name  : 'Colored Series',
+					labels: ['Jan', 'Feb','Mar', 'Apr', 'May', 'Jun'],
+					values: [20, 30, 10, 25, 15, 5]
+				}
+			],
+			{ x:0.5, y:0.6, w:'45%', h:3,
+				valAxisMaxVal:100,
+				barDir: 'bar',
+				valAxisMajorUnit: 50,
+				chartColors: ['0077BF','4E9D2D','ECAA00','5FC4E3','DE4216','154384'],
+				barGapWidthPct: 25
+			}
+		);
+
+		// TOP-RIGHT
+		slide.addChart(
+			pptx.charts.BAR,
+			[
+				{
+					name  : 'Too Many Colors Series',
+					labels: ['Jan', 'Feb','Mar', 'Apr', 'May', 'Jun'],
+					values: [.20, .30, .10, .25, .15, .05]
+				}
+			],
+			{ x:7, y:0.6, w:'45%', h:3,
+				valAxisMaxVal:1,
+				barDir: 'bar',
+				showValue: true,
+				dataLabelPosition: 'outEnd',
+				dataLabelFormatCode: '#%',
+				valAxisLabelFormatCode: '#%',
+				valAxisMajorUnit: 0.2,
+				chartColors: ['0077BF','4E9D2D','ECAA00','5FC4E3','DE4216','154384', '7D666A','A3C961','EF907B','9BA0A3'],
+				barGapWidthPct: 25
+			}
+		);
+
+		// BOTTOM-LEFT
+		slide.addChart(
+			pptx.charts.BAR,
+			[
+				{
+					name  : 'Two Color Series',
+					labels: ['Jan', 'Feb','Mar', 'Apr', 'May', 'Jun'],
+					values: [.20, .30, .10, .25, .15, .05]
+				}
+			],
+			{  x:0.5, y:3.6, w:'45%', h:3,
+				valAxisMaxVal:1,
+				barDir: 'bar',
+				showValue: true,
+				dataLabelPosition: 'outEnd',
+				dataLabelFormatCode: '#%',
+				valAxisLabelFormatCode: '0.#0',
+				chartColors: ['DE4216','154384'],
+				barGapWidthPct: 25
+			}
+		);
+
+		// BOTTOM-RIGHT
+		slide.addChart(
+			pptx.charts.BAR,
+			[
+				{
+					name  : 'Many values',
+					labels: 'A,B,C,D,E,F,G,H,I,J,K,L,M,N,O,P'.split(','),
+					values: [1.20, 2.30, 3.10, 4.25, 2.15, 6.05, 8.01, 2.02, 9.9, 0.9, 2.2, 3.33, 6.66, 7.77, 8.88]
+				}
+			], {  x:7, y:3.6, w:'45%', h:3,
+				valAxisMaxVal:10,
+				barDir: 'bar',
+				showValue: true,
+				dataLabelPosition: 'outEnd',
+				chartColors: ['0077BF','4E9D2D','ECAA00','5FC4E3','DE4216','154384', '7D666A','A3C961','EF907B','9BA0A3'],
+				barGapWidthPct: 25,
+				valAxisMajorUnit: 1
+			}
+		);
+	}
+
+	// SLIDE 6: Tornado Chart -------------------------------------------------------------
+	{
+		var slide = pptx.addNewSlide();
+		slide.addTable( [ [{ text:'Tornado Chart - Grid and Axis Formatting', options:gOptsTitle }] ], { x:0.5, y:0.13, w:12.5 } );
+
+		slide.addChart(
+			pptx.charts.BAR,
+			[
+				{
+					name: 'High',
+					labels: ['London', 'Munich', 'Tokyo'],
+					values: [.20, .32, .41]
+				},
+				{
+					name: 'Low',
+					labels: ['London', 'Munich', 'Tokyo'],
+					values: [-0.11, -0.22, -0.29]
+				}
+			],
+			{
+				x:0.5, y:0.5, w:'90%', h:'90%',
+				valAxisMaxVal:1,
+				barDir: 'bar',
+				axisLabelFormatCode: '#%',
+				gridLineColor: 'D8D8D8',
+				axisLineColor: 'D8D8D8',
+				catAxisLineShow: false,
+				valAxisLineShow: false,
+				barGrouping: 'stacked',
+				catAxisLabelPos: 'low',
+				valueBarColors: true,
+				//dataLabelPosition: 'outEnd',
+				chartColors: ['0077BF','4E9D2D','ECAA00','5FC4E3','DE4216','154384','7D666A','A3C961','EF907B','9BA0A3'],
+				invertedColors: ['0065A2','428526','C99100','51A7C1','BD3813','123970','6A575A','8BAB52','CB7A69','84888B'],
+				barGapWidthPct: 25,
+				valAxisMajorUnit: 0.2
+			}
+		);
+	}
+
+	// SLIDE 7: Line Chart: Line Smoothing, Line Size, Symbol Size -------------------------
+	{
+		var slide = pptx.addNewSlide();
+		slide.addTable( [ [{ text:'Chart Examples: Line Smoothing, Line Size, Line Shadow, Symbol Size', options:gOptsTitle }] ], { x:0.5, y:0.13, w:12.5 } );
 
 		slide.addText( '..', { x:0.5, y:0.6, w:6.0, h:3.0, fill:'F1F1F1', color:'F1F1F1'} );
 		var optsChartLine1 = { x:0.5, y:0.6, w:6.0, h:3.0,
@@ -781,19 +1055,22 @@ function genSlides_Chart(pptx) {
 		var optsChartLine1 = { x:0.5, y:4.0, w:6.0, h:3.0,
 			chartColors: [ COLOR_RED, COLOR_AMB, COLOR_GRN, COLOR_UNK ],
 			lineDataSymbolSize: 10,
+			lineShadow: 'none',
 			showLegend: true, legendPos: 'l'
 		};
 		slide.addChart( pptx.charts.LINE, arrDataLineStat, optsChartLine1 );
 
+		var shadowOpts = { type:'outer', color:'cd0011', blur:3, offset:12, angle:75, opacity:0.8 };
 		var optsChartLine2 = { x:7.0, y:4.0, w:6.0, h:3.0,
 			chartColors: [ COLOR_RED, COLOR_AMB, COLOR_GRN, COLOR_UNK ],
 			lineDataSymbolSize: 20,
+			lineShadow: shadowOpts,
 			showLegend: true, legendPos: 'b'
 		};
 		slide.addChart( pptx.charts.LINE, arrDataLineStat, optsChartLine2 );
 	}
 
-	// SLIDE 5: Line Chart: TEST: `lineDataSymbol` + `lineDataSymbolSize`
+	// SLIDE 8: Line Chart: TEST: `lineDataSymbol` + `lineDataSymbolSize` ------------------
 	{
 		var intWgap = 4.25;
 		var opts_lineDataSymbol = ['circle','dash','diamond','dot','none','square','triangle'];
@@ -814,7 +1091,7 @@ function genSlides_Chart(pptx) {
 		});
 	}
 
-	// SLIDE 6: Line Chart: Lots of Cats --------------------------------------------------
+	// SLIDE 9: Line Chart: Lots of Cats ---------------------------------------------------
 	{
 		var slide = pptx.addNewSlide();
 		slide.addTable( [ [{ text:'Chart Examples: Line Chart: Lots of Lines', options:gOptsTitle }] ], { x:0.5, y:0.13, w:12.5 } );
@@ -848,7 +1125,7 @@ function genSlides_Chart(pptx) {
 		slide.addChart( pptx.charts.LINE, arrDataTimeline, optsChartLine1 );
 	}
 
-	// SLIDE 7: Area Chart: Misc -------------------------------------------
+	// SLIDE 10: Area Chart: Misc -----------------------------------------------------------
 	{
 		var slide = pptx.addNewSlide();
 		slide.addTable( [ [{ text:'Chart Examples: Area Chart', options:gOptsTitle }] ], { x:0.5, y:0.13, w:12.5 } );
@@ -894,7 +1171,7 @@ function genSlides_Chart(pptx) {
 		slide.addChart( pptx.charts.AREA, arrDataTimeline2ser, optsChartLine4 );
 	}
 
-	// SLIDE 8: Pie Charts: All 4 Legend Options -------------------------------------------
+	// SLIDE 11: Pie Charts: All 4 Legend Options ------------------------------------------
 	{
 		var slide = pptx.addNewSlide();
 		slide.addTable( [ [{ text:'Chart Examples: Pie Charts: Legends', options:gOptsTitle }] ], { x:0.5, y:0.13, w:12.5 } );
@@ -927,10 +1204,10 @@ function genSlides_Chart(pptx) {
 		slide.addChart( pptx.charts.PIE, dataChartPieLocs, {x:9.8, y:4.0, w:3.2, h:3.2, dataBorder:{pt:'1',color:'F1F1F1'}, showLegend:true, legendPos:'b', showTitle:true, title:'Title & Legend'} );
 	}
 
-	// SLIDE 9: Pie Chart -----------------------------------------------------------------
+	// SLIDE 12: Doughnut Chart ------------------------------------------------------------
 	{
 		var slide = pptx.addNewSlide();
-		slide.addTable( [ [{ text:'Chart Examples: Pie Chart', options:gOptsTitle }] ], { x:0.5, y:0.13, w:12.5 } );
+		slide.addTable( [ [{ text:'Chart Examples: Doughnut Chart', options:gOptsTitle }] ], { x:0.5, y:0.13, w:12.5 } );
 
 		var optsChartPie1 = {
 			x:0.5, y:1.0, w:6.0, h:6.0,
@@ -938,7 +1215,6 @@ function genSlides_Chart(pptx) {
 			dataBorder       : { pt:'2', color:'F1F1F1' },
 			dataLabelColor   : 'FFFFFF',
 			dataLabelFontSize: 14,
-			//dataLabelPosition: 'ctr',
 
 			legendPos : 'r',
 
@@ -948,13 +1224,15 @@ function genSlides_Chart(pptx) {
 			showLegend : true,
 			showTitle  : false,
 
+			holeSize: 70,
+
 			title        : 'Project Status',
 			titleColor   : '33CF22',
 			titleFontFace: 'Helvetica Neue',
 			titleFontSize: 24
 		};
 		slide.addText( '.', {x:0.5, y:1.0, w:6.0, h:6.0, fill:'F1F1F1', color:'F1F1F1'} );
-		slide.addChart(pptx.charts.PIE, dataChartPieStat, optsChartPie1 );
+		slide.addChart(pptx.charts.DOUGHNUT, dataChartPieStat, optsChartPie1 );
 
 		var optsChartPie2 = {
 			x:7.0, y:1.0, w:6, h:6,
@@ -967,7 +1245,7 @@ function genSlides_Chart(pptx) {
 			showTitle  : false,
 			title: 'Resource Totals by Location'
 		};
-		slide.addChart(pptx.charts.PIE, dataChartPieLocs, optsChartPie2 );
+		slide.addChart(pptx.charts.DOUGHNUT, dataChartPieLocs, optsChartPie2 );
 	}
 }
 
@@ -1158,8 +1436,8 @@ function genSlides_Text(pptx) {
 		slide.addText("Line-Spacing (bullets):", { x:7.5, y:4.45, w:'40%', h:0.38, color:'0088CC' });
 		slide.addText(
 			[
-		      { text:'Line Spacing\n35pt', options:{ font_size:24, bullet:true, color:'99ABCC', lineSpacing:35 } }
-		    ],
+			  { text:'Line Spacing\n35pt', options:{ font_size:24, bullet:true, color:'99ABCC', lineSpacing:35 } }
+			],
 			{ x:7.5, y:4.85, w:5.25, h:1.15, margin:0.1, fill:'f1f1f1' }
 		);
 	}
@@ -1246,7 +1524,7 @@ function genSlides_Text(pptx) {
 				{ text:'Sub'},
 				{ text:'Subscript', options:{ subscript:true } },
 				{ text:' // Super'},
-			    { text:'Superscript', options:{ superscript:true } }
+				{ text:'Superscript', options:{ superscript:true } }
 			],
 			{ x:10, y:5.5 }
 		);